Remove message search from chat session view - just use command + f like a normal person
